apps/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
pytest==7.4.2
pytest-cov==4.1.0
httpx==0.24.1
supabase==1.0.4
python-jose==3.3.0 